requirement.txt
appdirs==1.4.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.6
channels==2.1.5
charset-normalizer==2.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
daphne==2.5.0
defusedxml==0.7.1
distlib==0.3.2
Django==2.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.0.12
GrabzIt==3.4.5.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
pipenv==2021.5.29
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pyOpenSSL==20.0.1
PyPDF2==1.26.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tqdm==4.58.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.4.7
virtualenv-clone==0.5.4
zope.interface==5.4.0
